// src/app/shared/models/country.model.ts
export interface CountryResponseDto {
  id: number
  name: string
  province: string
  city: string
  postalCode: string
  latitude: number
  longitude: number
}

export interface Country {
  id: number
  name: string
  province: string
  city: string
  postalCode: string
  latitude: number
  longitude: number
}